Fix for invalid Windows-built layer.

+++ build_lambda_artifacts.py
@@ -4,6 +4,10 @@ import subprocess
 import sys
 
 
+# if dependencies are not changes, it's useful to set this as false.
+skip_layer_building = bool(int(os.getenv('SKIP_LAYER_BUILDING')))
+
+
 def build_lambda_zip(src_folder_name):
     shutil.make_archive(f'artifacts/{src_folder_name}', 'zip', f'src/{src_folder_name}')
 
@@ -23,12 +27,28 @@ def build_layer_zip():
 
 
 if __name__ == '__main__':
-    shutil.rmtree('artifacts', ignore_errors=True)
-    shutil.rmtree('installed_dependencies', ignore_errors=True)
-    os.mkdir('artifacts')
+    if not os.path.exists('artifacts'):
+        os.mkdir('artifacts')
 
-    install_layer_dependencies()
-    build_layer_zip()
+    if not skip_layer_building:
+        print('Building lambda layer...')
+        if os.path.exists('installed_dependencies'):
+            shutil.rmtree('installed_dependencies', ignore_errors=True)
+        if os.path.exists('artifacts/lambda_layer.zip'):
+            os.remove('artifacts/lambda_layer.zip')
+        install_layer_dependencies()
+        build_layer_zip()
+        print('Lambda layer ZIP is created.')
+    else:
+        print('Layer building is skipped.')
+
+    if os.path.exists('artifacts/discord_interaction_lambda.zip'):
+        os.remove('artifacts/discord_interaction_lambda.zip')
     build_lambda_zip('discord_interaction_lambda')
+    print('Discord interaction lambda ZIP is created.')
+
+    if os.path.exists('artifacts/scheduled_lambda.zip'):
+        os.remove('artifacts/scheduled_lambda.zip')
     build_lambda_zip('scheduled_lambda')
+    print('Scheduled lambda ZIP is created.')
